requirements.txt
numpy
funlib.learn.tensorflow
